This adds the implemenation of MSAL to the iOS graph notifications sample.
This commit will show the MSAL sign in process with both AAD and MSA to get an MSAL oAuth2 token, the initialization of notifications manager, and the subscription of the application to MSGraph using the received MSAL oAuth2 Token, and the AppPackageName to be put into secrets.h.

This change marks feature completeness of the iOS sample application.